Se inician pruebas unitarias

+++ src/main/java/com/natamarin/mutants/repository/MutantRepository.java
@@ -11,5 +11,5 @@ import com.natamarin.mutants.entity.AdnVerificado;
  */
 public interface MutantRepository extends CrudRepository<AdnVerificado, Long> {
 
-	Long countByEsMutante(Boolean esMutante);
+	int countByEsMutante(Boolean esMutante);
 }
